pyproject.toml
[project]
name = "ai-agents"
version = "0.1.0"
description = "AI agent fleet — devops, finance, media, office, social. Shared loop, domain-specific tools."
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.25.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/shared", "src/agents"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
